feat(tabs): add onPress callback (#534)
* feat(tabs): add onTap function

This function is triggered on click.

* refactor(tabs): Rename onTap callback to onPress

Renamed onTap to onPress for a more accurate description of its behavior.

Co-Authored-By: Pante <matthiasngeo@gmail.com>

* test(tabs): add test case to verify onPress event is triggered when tapping a tab

Added a new test case to ensure the onPress event is correctly triggered when a tab is tapped, and to verify the controller index and content update accordingly.

* docs(CHANGELOG): Update CHANGELOG.md to record the addition of FTabs.onPress

Updated the CHANGELOG.md file, adding an explanation about the new onPress property of the FTabs component. This callback is triggered when a tab is pressed and is called before the tab switch animation starts and before the controller is updated.

* Commit from GitHub Actions (Forui Presubmit)

---------

Co-authored-by: Pante <matthiasngeo@gmail.com>
Co-authored-by: Pante <Pante@users.noreply.github.com>