Unlink samples from root workspace
auto_route's maintenance has been lagging behind for awhile now. It's been causing issues with forui_internal_gen as auto_route_generator somehow has a direct dependency on lean_builder, which in turn depends on an old version of _fe_analyzer_shared.

This is particularly bad since that was never intended for external use.

In short, I've lost faith in the maintenance of auto_route and we should off-board ASAP.