Improve dot-shorthand support (#756)
* Update forui to use & support dot-shorthand

* Update forui_hooks

* Unlink samples from root workspace

auto_route's maintenance has been lagging behind for awhile now. It's been causing issues with forui_internal_gen as auto_route_generator somehow has a direct dependency on lean_builder, which in turn depends on an old version of _fe_analyzer_shared.

This is particularly bad since that was never intended for external use.

In short, I've lost faith in the maintenance of auto_route and we should off-board ASAP.

* Tweak toaster animation curve

* WIP

* Tweak FPortalSpacing

* Upgrade docs dependency

* Fix issues with samples

* Prepare Forui Hooks for review

* Prepare Samples for review

* Prepare Forui Internal Gen for review

* Update GH actions

* Fix CLI snippets

* Prepare Forui for review

* Bump example project version

---------

Co-authored-by: Pante <Pante@users.noreply.github.com>